+++ main.py
@@ -3,7 +3,7 @@ import pygame
 import json
 
 # Other code files
-import rules.rule_ui as r_ui
+import universal.side_bar as sb
 import fields.fields as fields
 import fields.fcontainer as fc
 import player.player as pl
@@ -75,9 +75,9 @@ def afterTurn(player):
         else:
             st_transact.payRent(player, fc.f_container[player.fid])
             
-    elif fc.f_container[player1.fid].type == "investment":
-        if fc.f_container[player1.fid].owner == "Bank":
-            invest_transact.invest(player1, fc.f_container[player.fid])
+    elif fc.f_container[player.fid].type == "investment":
+        if fc.f_container[player.fid].owner == "Bank":
+            invest_transact.invest(player, fc.f_container[player.fid])
             
         elif fc.f_container[player.fid].owner != player:
             invest_transact.earn_money(player, fc.f_container[player.fid])
@@ -150,7 +150,7 @@ while running:
     screen.fill("purple")
 
     # RENDER YOUR GAME HERE
-    r_ui.rule_ui_setup(screen)
+    sb.sb_setup(screen)
     
     for fld in fc.f_container:
         fld.field_placement(screen)