+++ transactions/invest_transact.py
@@ -1,7 +1,7 @@
 def invest(player, investment):
     if player.balance >= investment.price:
         player.balance -= investment.price
-        investment.owner == player
+        investment.owner = player
         
 def earn_money(player, investment):
     player.balance -= investment.rent * 4