+++ CHANGELOG.md
@@ -8,7 +8,10 @@ This build is not intended for submission, but rather for the developers to keep
 
 - Some code has been restructured.
 - That means a fix for the player logic had to be made.
+- The rule UI file is now the sidebar file in `universal`.
 - The players can move again.
+- The ownership issues have been fixed.
+- The investment issues have been fixed.
 - The dynamic rules are in development.
     - The house and hotel prices are now variables which can be changed.
     - The events "Shop closure", "Housing crisis" and "Housing abundance" have been prepared.